Move all this to ArcWeapon
