+++ code/Weapons/Base/PowerArrow.cs
@@ -1,6 +1,6 @@
 ﻿using Sandbox;
 
-namespace TerryForm.Pawn
+namespace TerryForm.Weapons
 {
 	public partial class PowerArrow : RenderEntity
 	{
@@ -18,10 +18,10 @@ namespace TerryForm.Pawn
 			var vertexBuffer = Render.GetDynamicVB( true );
 
 			// Line
-			Vertex a = new( startPos - size, Vector3.Up, Vector3.Right, new Vector4( 0, 1, 0, 0 ) );
-			Vertex b = new( startPos + size, Vector3.Up, Vector3.Right, new Vector4( 1, 1, 0, 0 ) );
-			Vertex c = new( endPos + size, Vector3.Up, Vector3.Right, new Vector4( 1, 0, 0, 0 ) );
-			Vertex d = new( endPos - size, Vector3.Up, Vector3.Right, new Vector4( 0, 0, 0, 0 ) );
+			Vertex a = new( startPos - size, -Vector3.Forward, Vector3.Right, new Vector4( 0, 1, 0, 0 ) );
+			Vertex b = new( startPos + size, -Vector3.Forward, Vector3.Right, new Vector4( 1, 1, 0, 0 ) );
+			Vertex c = new( endPos + size, -Vector3.Forward, Vector3.Right, new Vector4( 1, 0, 0, 0 ) );
+			Vertex d = new( endPos - size, -Vector3.Forward, Vector3.Right, new Vector4( 0, 0, 0, 0 ) );
 
 			vertexBuffer.Add( a );
 			vertexBuffer.Add( b );
@@ -32,9 +32,9 @@ namespace TerryForm.Pawn
 			vertexBuffer.AddTriangleIndex( 2, 1, 4 );
 
 			// Add the arrow tip
-			Vertex e = new( endPos + size * 1.75f, Vector3.Up, Vector3.Right, new Vector4( 1, 0, 0, 0 ) );
-			Vertex f = new( endPos - size * 1.75f, Vector3.Up, Vector3.Right, new Vector4( 0, 0, 0, 0 ) );
-			Vertex g = new( endPos + direction * 8, Vector3.Up, Vector3.Right, new Vector4( 1, 0, 0, 0 ) );
+			Vertex e = new( endPos + size * 1.75f, -Vector3.Forward, Vector3.Right, new Vector4( 1, 0, 0, 0 ) );
+			Vertex f = new( endPos - size * 1.75f, -Vector3.Forward, Vector3.Right, new Vector4( 0, 0, 0, 0 ) );
+			Vertex g = new( endPos + direction * 8, -Vector3.Forward, Vector3.Right, new Vector4( 1, 0, 0, 0 ) );
 
 			vertexBuffer.Add( e );
 			vertexBuffer.Add( f );
@@ -58,7 +58,7 @@ namespace TerryForm.Pawn
 			var size = Vector3.Cross( Direction, Vector3.Up ) * 3f;
 
 			//var color = ColorConvert.HSLToRGB( 120 - (int)(Power * Power * 120), 1.0f, 0.5f );
-			DrawArrow( obj, startPos, endPos, Direction, size, Color.White );
+			DrawArrow( obj, startPos, endPos, Direction, size, Color.Red );
 		}
 	}
 }
